+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `set` method on `collSpecable`;
 
+### Fix
+
+- Ensure promise is always returned from `activate` method;
+
 ---
 
 ## [0.0.1] - 2021-03-30